+++ src/ui/hooks/useStaticAsset.js
