+++ src/ui/states/approveTransaction/TransactionDetails.js
