+++ src/ui/states/approveTransaction/approve.transaction.css
