+++ src/ui/states/approveTransaction/details/DocumentCreateTransition.js
