+++ src/ui/states/approveTransaction/details/DocumentsBatchDetails.js
