+++ src/ui/states/approveTransaction/details/documents.batch.details.css
