--- src/ui/style.pcss
@@ -1,1 +1,0 @@
-@import "../../node_modules/tailwindcss";
