+++ src/ui/styles/app.pcss
@@ -1,3 +1,5 @@
+@import "tailwindcss";
+
 body {
     margin: 0;
     display: flex;
